@@ -3,38 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.22.8",
-    "@babel/runtime": "7.0.0-beta.42",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.3",
-    "@mui/icons-material": "^5.11.11",
-    "@mui/material": "^5.11.12",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/base": "^5.0.0-beta.8",
+    "@mui/icons-material": "^5.14.1",
+    "@mui/material": "^5.14.2",
+    "@mui/styles": "^5.14.1",
     "@react-oauth/google": "^0.9.0",
-    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
+    "install": "^0.13.0",
     "js-cookie": "^3.0.5",
     "jwt-decode": "^3.1.2",
-    "material-icons": "^1.13.2",
-    "material-ui": "^1.0.0-beta.47",
-    "prop-types": "^15.8.1",
-    "react": "18.1.0",
+    "npm": "^9.8.1",
+    "react": "^18.2.0",
     "react-axios": "^2.0.6",
     "react-d3-tree": "^3.5.1",
-    "react-dnd": "11.1.3",
-    "react-dnd-html5-backend": "11.1.3",
-    "react-dom": "^18.1.0",
+    "react-dom": "^18.2.0",
     "react-google-login": "^5.2.2",
-    "react-organizational-chart": "1.1.0",
-    "react-reveal": "^1.2.2",
-    "react-router-dom": "^6.0.0",
-    "react-scripts": "3.0.1",
-    "react-scroll": "^1.7.9",
-    "react-slick": "^0.23.1",
-    "react-transition-group": "^4.4.5",
+    "react-router-dom": "^6.14.2",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
