@@ -1,12 +1,10 @@
 import React from "react";
-import { ThemeProvider } from "@material-ui/core";
+import { ThemeProvider } from "@mui/styles";
 import GlobalStyles from "./components/GlobalStyles";
 import theme from "./theme";
 import AppRoutes from "./routes/AppRoutes";
 
 const App = () => {
-  //const routing = useRoutes(Array.isArray(AppRoutes) ? AppRoutes : []);
-
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyles />
