@@ -1,7 +1,7 @@
 import React, { useLayoutEffect } from "react";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
-import { Grid } from "@material-ui/core";
+import  Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import Table from "@mui/material/Table";
 import TableBody from "@mui/material/TableBody";
@@ -12,14 +12,14 @@ import TableHead from "@mui/material/TableHead";
 import CloseIcon from "@mui/icons-material/Close";
 import TableRow from "@mui/material/TableRow";
 import Paper from "@mui/material/Paper";
-import { Button } from "@material-ui/core";
+import  Button  from "@mui/material/Button";
 import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined";
-import IconButton from "material-ui/IconButton";
+import IconButton from "@mui/material/IconButton";
 import PersonAddAltOutlinedIcon from "@mui/icons-material/PersonAddAltOutlined";
 import EmailIcon from "@mui/icons-material/Email";
-import TextField from "material-ui/TextField";
+import TextField from "@mui/material/TextField";
 import "./stylesAdminUsu.css";
-import { Dialog, ListItem } from "material-ui";
+import { Dialog, ListItem } from "@mui/material";
 import List from "@mui/material/List";
 import { getDeleteUser, getAllUsers, getSaveNewUser } from "../../api/usuarios";
 import Navbar from "../../components/NavBar/Navbar";
