@@ -1,5 +1,5 @@
-import { createTheme } from "@mui/material/styles";
-import { colors } from '@material-ui/core';
+import {createTheme}  from '@mui/material/styles';
+import { colors } from '@mui/material';
 import shadows from './shadows';
 import typography from './typography';
 
