@@ -1,4 +1,4 @@
-import IconButton from "material-ui/IconButton";
+import IconButton from "@mui/material/IconButton";
 import MenuIcon from "@mui/icons-material/Menu";
 import InputOutlinedIcon from "@mui/icons-material/InputOutlined";
 import * as React from "react";
